\name{describe.gamma.gee}
\alias{describe.gamma.gee}
\title{Describe the \code{gamma.gee} model to Zelig}
\usage{
  \method{describe}{gamma.gee}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{gamma.gee} model to Zelig
}
\author{
  Patrick Lam
}

