\name{describe.normal.gee}
\alias{describe.normal.gee}
\title{Describe the \code{normal.gee} model to Zelig}
\usage{
  \method{describe}{normal.gee}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{normal.gee} model to Zelig
}
\author{
  Patrick Lam
}

