\name{gee.zelig-package}
\title{gee...}
\description{gee.zelig}
\details{\tabular{ll}{
Package: \tab gee.zelig\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{gee.zelig-package}
\alias{gee.zelig}
\docType{package}
\author{Patrick Lam}
\keyword{package}
