\name{param.logit.gee}
\alias{param.logit.gee}
\title{param method for the ``logit.gee'' Zelig model
Return parameter estimates for the ``logit'' GLM in Zelig.}
\usage{
  \method{param}{logit.gee}(obj, num, ...)
}
\arguments{
  \item{obj}{a `zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a `parameters' object
}
\description{
  param method for the ``logit.gee'' Zelig model Return
  parameter estimates for the ``logit'' GLM in Zelig.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

