\name{plot.sim.normal.gee}
\alias{plot.sim.normal.gee}
\title{Plot Simualted Quantities of Interest for GEE Zelig Models...}
\usage{\method{plot}{sim.normal.gee}(x, ...)}
\description{Plot Simualted Quantities of Interest for GEE Zelig Models}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a simulated GEE object containing simulations of quantities of
interest}
\item{...}{ignored parameters}}
