\name{qi.probit.gee}
\alias{qi.probit.gee}
\title{Compute Quantities of Interest for the Zelig Model probit...}
\description{Compute Quantities of Interest for the Zelig Model probit.gee}
\value{a list of key-value pairs specifying pairing titles of quantities of interest
with their simulations}
\arguments{\item{obj}{a zelig object}
\item{x}{a setx object}
\item{x1}{an optional setx object}
\item{y}{...}
\item{num}{an integer specifying the number of simulations to compute}
\item{param}{a parameters object}}
