\name{vcov.gee.naive}
\alias{vcov.gee.naive}
\title{Covariance Matrix for Naive GEE...}
\usage{\method{vcov}{gee.naive}(object, ...)}
\description{Covariance Matrix for Naive GEE}
\value{the naive variance-covariance matrix}
\author{Skyler}
\arguments{\item{object}{a \code{gee.naive} object}
\item{...}{ignored parameters}}
