\name{vcov.gee.robust}
\alias{vcov.gee.robust}
\title{Covariance Matrix for Robust GEE}
\usage{
  \method{vcov}{gee.robust}(object, ...)
}
\arguments{
  \item{object}{a \code{gee.robust} object}

  \item{...}{ignored parameters}
}
\value{
  the robust variance-covariance matrix
}
\description{
  Covariance Matrix for Robust GEE
}
\author{
  Skyler
}

