\name{zelig2poisson.gee}
\alias{zelig2poisson.gee}
\title{Interface between the Zelig Model poisson...}
\usage{zelig2poisson.gee(formula, id, robust, ..., R, corstr="independence",
    data)}
\description{Interface between the Zelig Model poisson.gee and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
