\name{zelig2probit.gee}
\alias{zelig2probit.gee}
\title{Interface between the Zelig Model probit...}
\usage{zelig2probit.gee(formula, id, robust, ..., R, corstr="independence",
    data)}
\description{Interface between the Zelig Model probit.gee and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
