\docType{package}
\name{ZeligGEE-package}
\alias{ZeligGEE}
\alias{ZeligGEE-package}
\title{ZeligGEE}
\description{
  \tabular{ll}{ Package: \tab ZeligGEE\cr Version: \tab
  0.2\cr Date: \tab 2012-09-19\cr Depends: Zelig (>=
  4.0-6), gee\cr License: \tab GPL version 2 or newer\cr }
}
\details{
  Edit this description
}
\author{
  Patrick Lam
}
\keyword{package}

