\name{clean.up.gamma.gee}
\alias{clean.up.gamma.gee}
\title{Remove Negative Simulations from Gamma GEE Parameter Simulations...}
\usage{clean.up.gamma.gee(object, x, x1, bootstrap=FALSE, bootfn, data, param,
    num=1000)}
\description{Remove Negative Simulations from Gamma GEE Parameter Simulations}
\arguments{\item{obj}{}
\item{x}{}
\item{x1}{}}
