\name{param.gamma.gee}
\alias{param.gamma.gee}
\title{param method for the ``gamma.gee'' Zelig model
Return parameter estimates for the ``gamma'' GLM in Zelig.}
\usage{
  \method{param}{gamma.gee}(obj, num, ...)
}
\arguments{
  \item{obj}{a `zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a `parameters' object
}
\description{
  param method for the ``gamma.gee'' Zelig model Return
  parameter estimates for the ``gamma'' GLM in Zelig.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

