\name{robust.hook}
\alias{robust.hook}
\title{Classify Fitted Object as Naive or Robust...}
\usage{robust.hook(obj, Zall, Call, robust, ...)}
\description{Classify Fitted Object as Naive or Robust}
\details{This hook is ran after the call to the external mode. It sets the class of
the object (in addition to its other designations) as 'gee.naive' or 
'gee.robust' depending on the value of the \code{robust} parameter.}
\value{4}
\author{Skyler}
\arguments{\item{1}{}
\item{2}{}
\item{3}{}}
