\name{vcov.gee.naive}
\alias{vcov.gee.naive}
\title{Covariance Matrix for Naive GEE}
\usage{
  \method{vcov}{gee.naive}(object, ...)
}
\arguments{
  \item{object}{a \code{gee.naive} object}

  \item{...}{ignored parameters}
}
\value{
  the naive variance-covariance matrix
}
\description{
  Covariance Matrix for Naive GEE
}
\author{
  Skyler
}

