\name{vcov.gee.robust}
\alias{vcov.gee.robust}
\title{Covariance Matrix for Robust GEE...}
\usage{\method{vcov}{gee.robust}(object, ...)}
\description{Covariance Matrix for Robust GEE}
\value{the robust variance-covariance matrix}
\author{Skyler}
\arguments{\item{object}{a \code{gee.robust} object}
\item{...}{ignored parameters}}
